+++ modules/geoip/moduleconfig.h
@@ -34,7 +34,8 @@
 #endif // __DATE__
 
 // metamod plugin?
-#define USE_METAMOD
+// KTP: Disabled for extension mode
+// #define USE_METAMOD
 
 // use memory manager/tester?
 // note that if you use this, you cannot construct/allocate